Added *, /, ** operations to bitstring
Multiplication of bitstrings is GF(2) polynomial multiplication mod x^l-1, where l is the common length of the bitstrings. The constant term of the polynomial corresponds to the first bit of the bitstring (index 0) which is the most significant bit of the int value of the bitstring. As x^l-1 is not irreducible (except for l==1), attempts to divide by a bitstring (i.e., polynomial) not coprime with x^l-1 will raise a ZeroDivisionError, whether by explicit division or negative exponent.